requirements.txt
azure-functions
fastapi
uvicorn
requests
beautifulsoup4
pandas
python-multipart
# streamlit # Optional: kept for legacy app.py compatibility if needed, but not for Azure
